--- Car_Prediction/net.py
@@ -1,8 +1,0 @@
-import abc
-
-
-class Net(metaclass=abc.ABCMeta):
-
-    @abc.abstractmethod
-    def setup_model(self):
-        return
